Feature: Support Parallel Cursor
With the segment number increasing, the performance bottleneck on master
node becomes more and more obvious.

The parallel cursor feature is designed to reduce the heavy burdens of
master node. If possible it will not gather the result to master, and
redirect the result to segments. When it is executed, user can setup
retrieve mode connection to segments for retrieving result data
parallelly.

This feature is originally designed for implementing gp2gp feature, but it
can also be used for any tools need to interact with gpdb for heavy
workload.

Co-Authored-By: Adam Lee <ali@pivotal.io>
Co-Authored-By: Chen Mulong <chenmulong@gmail.com>
Co-Authored-By: (Jerome)Junfeng Yang <jeyang@pivotal.io>
Co-Authored-By: Max Yang <myang@pivotal.io>
Co-Authored-By: Ming LI <mli@apache.org>
Co-Authored-By: Tingfang Bao <bbao@pivotal.io>
Co-Authored-By: Wen Lin <wlin@pivotal.io>