Fixed flaky test
The endpoint on seg0 for c1 should firstly become to RELEASED (the retrieve process set it),
and then was removed (during the endpoint QE cancelled), so check the
status is RELEASED make this test case flaky.

The solution is to sleep for a while, so that the cancel interrupt can
be invoked, which will remove the endpoint info.