Keep token the same in the same sessions
Since the token has been changed to be generated on the QE, the old
static trick won't work anymore to keep the unique token in the session.
Also, we don't have to worry about the latch & session abort issues
anymore since the latch has been removed, we don't have to reset the
token every time when endpoint created.

Authored-by: Chen Mulong <muchen@pivotal.io>